Fix bugs found while reading the code and testing against hardware
Most of these were found by reading, a few only showed up with a radio
plugged in.

Crashes and lifetimes:

* RadioFactory::ListDevices built the USBRadioFactory as a stack local, so
  its destructor ran libusb_exit before any device could be opened. Every
  USB operation segfaulted: --info, --get-status, --get-time, --dump-reg
  and --flash. Only --list survived, because it never opens anything. The
  factory now outlives the handles it hands out.
* The lambda that opens a device captured a pointer that had already gone.
* The libusb context leaked on the error paths.

Checks that could not fail:

* IsCompatible compared typeid of the same expression in five handlers, so
  it was always true. Any firmware file was accepted for any radio.
* The H8SX sum check joined four conditions with && instead of ||.
* The Ailunce check ended in || true.
* YaesuFW::SupportsRadioModel returned true for every model.

Wrong output:

* DFUStatusReport streamed a uint8_t as a character, so a status line ended
  in a raw NUL and made the output a binary file to grep.
* FormatBytes had no MiB case, PrintHex rotated its output, and XORTool
  overflowed a uint8_t.
* --wrap with an SGL radio failed outright.
* The SGL error message was built from a stringstream that had already been
  moved from.

Tests:

* test_util.cpp checked Fletcher-16 against 0x0627, which is not the
  checksum of the vector it uses. Corrected to 0xDEBE.
* Added round trip tests over the firmware containers.

+++ include/radio_tool/dfu/dfu.hpp
@@ -217,7 +217,10 @@ namespace radio_tool::dfu
             out << "Status: " << radio_tool::dfu::ToString(status) << ", "
                 << "Timeout: 0x" << std::setfill('0') << std::setw(2) << std::hex << timeout << ", "
                 << "State: " << radio_tool::dfu::ToString(state) << ", "
-                << "Discarded: 0x" << std::setfill('0') << std::setw(2) << std::hex << discarded;
+                //cast: discarded is a byte, and streaming it directly printed
+                //it as a character, so a status line ended in a raw NUL
+                << "Discarded: 0x" << std::setfill('0') << std::setw(2) << std::hex
+                << static_cast<unsigned int>(discarded);
             return out.str();
         }
